Remove WhatDoTheyKnow namespace
This class is used in all installs of Alaveteli
